requirements.txt

numpy
scipy
torch
torchvision
matplotlib
pandas
pyyaml
